refactor: table / order JPA 반영
