+++ src/main/java/kitchenpos/menu/repository/MenuGroupRepository.java
@@ -1,6 +1,5 @@
 package kitchenpos.menu.repository;
 
-import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
