+++ src/main/java/kitchenpos/order/constant/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.constant;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
