+++ src/main/java/kitchenpos/order/repository/OrderLineItemRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.order.repository;
+
+import kitchenpos.order.domain.OrderLineItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
+}
