+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.order.repository;
+
+import kitchenpos.order.domain.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
