+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -3,7 +3,7 @@ package kitchenpos.table.application;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.table.dao.OrderTableDao;
 import kitchenpos.table.dao.TableGroupDao;
-import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.constant.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import org.springframework.stereotype.Service;
