+++ src/main/java/kitchenpos/table/repository/OrderTableRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.table.repository;
+
+import kitchenpos.table.domain.OrderTable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
+}
